refactor: 타입 분리 및 import 수정
