+++ src/features/shared/types.ts
@@ -1,0 +1,7 @@
+/**
+ * 액션 상태 타입
+ */
+export type ActionState = {
+  status: boolean | null;
+  error?: string;
+};
